+++ core/test/java/MapBoardTests.java
@@ -1,6 +1,6 @@
 import it.units.sdm.project.board.Position;
 import it.units.sdm.project.board.Stone;
-import it.units.sdm.project.board.terminal.MapBoard;
+import it.units.sdm.project.board.MapBoard;
 import it.units.sdm.project.interfaces.Board;
 import com.badlogic.gdx.graphics.Color;
 import org.junit.jupiter.api.Assertions;
