+++ core/test/java/utility/TestUtilities.java
@@ -3,7 +3,7 @@ package utility;
 import com.badlogic.gdx.graphics.Color;
 import it.units.sdm.project.board.Position;
 import it.units.sdm.project.board.Stone;
-import it.units.sdm.project.board.terminal.MapBoard;
+import it.units.sdm.project.board.MapBoard;
 import it.units.sdm.project.interfaces.Board;
 import org.jetbrains.annotations.NotNull;
