+++ terminal/src/main/java/it/units/sdm/project/TerminalLauncher.java
@@ -1,7 +1,7 @@
 package it.units.sdm.project;
 
 import com.badlogic.gdx.graphics.Color;
-import it.units.sdm.project.board.terminal.MapBoard;
+import it.units.sdm.project.board.MapBoard;
 import it.units.sdm.project.board.Stone;
 import it.units.sdm.project.game.Player;
 import it.units.sdm.project.interfaces.Board;
